requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
boto3>=1.34.0
python-multipart>=0.0.6
alembic>=1.13.0
openai>=1.12.0
sentry-sdk[fastapi]>=1.40.0
